requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.1
colorama==0.4.6
frozenlist==1.6.0
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
multidict==6.4.3
paho-mqtt==2.1.0
propcache==0.3.1
pycryptodome==3.22.0
python-weather==2.0.7
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
tuya-connector-python==0.1.2
tuya-iot-py-sdk==0.6.6
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
yarl==1.20.0
logfmter==0.0.9
concurrent-log-handler==0.9.26